Updates Weaver tests to focus on thing as central entity
